logto/requirements.txt
aiohttp==3.8.5
pydantic==2.1.1
PyJWT==2.8.0
